requirements.txt
Flask==1.0.2
line-bot-sdk==1.8.0
numpy==1.14.3
pandas==0.23.0